+++ pkg/sass-parser/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.2.6
 
-* No user-visible changes.
+* Add `BooleanExpression` and `NumberExpression`.
 
 ## 0.2.5